# Fortran backend for 1D EM forward modelling and Numpy broadcasting