… simpegEM1D/RTEfun_vec.py
… simpegEM1D/Waveforms.py