@prisae 2018-08-31
Good catch on the overwriting of just created arrays. I saw that when I worked on the filters and wondered about it, but then left it aside.

You still have things like
```
lambd = np.empty([self.survey.frequency.size, n_filter], order='F')
lambd[:, :], _ = get_spline_values(self.fhtfilt, r, self.hankel_pts_per_dec)
```
Couldn't you just write
```
lambd, _ = get_spline_values(self.fhtfilt, r, self.hankel_pts_per_dec)
```

I might have time on the weekend to go a bit through your code, or then later in September. However, I probably wait until @sgkang went through. He knows the code in and out, I just had a look related to Hankel/Fourier transform.