@prisae 2018-08-31
One comment related to your question in #14 : In `empymod` the entire kernel is in python, where the calculation is done in a massive matrix that includes all frequencies, wavenumbers, and offsets (if RAM permits and depending on the Hankel/Fourier transform method; otherwise it might loop over frequencies or over offsets or both; but it always includes all wavenumbers).

I later profiled the kernel, and the most time consuming parts where the square roots and the exponential functions (`np.sqrt`, `np.exp`) which included this big matrix. I tried various things to check if I could speed it up, such as `numba`, `Cython`. However, not really successful. This led me to the conclusion that the most time consuming part of `empymod` (I think in the order of 90 % runtime), `np.sqrt` and `np.exp` calls, simply take that amount to calculate, and cannot be further optimized (if using `conda` with `mkl` etc). So rewriting the python code into a Cython loop over frequencies, offsets, and wavenumbers does not improve it, because the square root and the exponential in this (then compiled) loop will still take that time.

It would therefore be interesting to know if your Fortran implementation is really faster than a pure, vectorized python implementation (which is what #14 is about).

Of course, that is a luxury problem more out of interest than anything else, and probably not worth the extra work, as you have already done the Fortran version and it proves to be faster than the existing version.