@leonfoks 2018-08-31
First comment re:

```python
lambd = np.empty([self.survey.frequency.size, n_filter], order='F')
lambd[:, :], _ = get_spline_values(self.fhtfilt, r, self.hankel_pts_per_dec)
```

This was done on purpose. Since lambd is an input to the forward modeller, I had to make sure that its ordering was Fortran.  Otherwise, when a numpy array in order='C' is passed to a Fortran subroutine, there is an implicit copy due to the different row/column major ordering between C/Fortran. The first line instantiates the memory without specifying a value, so very fast. While the second line might return a C ordered numpy array, we can put it into Fortran order with no issues. This saves me from digging down into get_spline_values() and changing the ordering there.

If another part of the code base uses lambd, there should not be a dramatic change in efficiency for those Python codes that were once using lambd with C ordering.

Second comment re: 
I agree with what you are saying. There is little to no speed up gained when migrating a vectorized Numpy function to Fortran/C when there is a single Pythonic loop.  By this I say a pythonic loop is 

```python
for i in range(nFrquencies):
   Do a bunch of vectorized Numpy stuff
```

The speed is equivalent between Numpy broadcasting vs loops in Fortran/C. Since they are doing the same operations on the backend, the time should not change. The downside with Numpy broadcasting is the extra memory (sometimes a lot!) to carryout the operations. So I only expected Fortran to be a little faster than the vectorized Numpy implementation for the sensitivity matrix. The gains were mainly the memory, and removing the need to keep instantiating 2D and 3D blocks of memory with every call to the sensitivity matrix. Allocation of memory takes time!

If the code was written without Numpy (i.e. pythonic) there absolutely are efficiency gains.

However, and I found this out yesterday, when the Fortran extension is added to the setup.py file, and we "pip install simpegEM1D", f2py is called which creates the glue code to the Fortran module.  Pip then compiles the module using some default compile flags namely '-Ox -g', where '-O' is the optimization level, and '-g' adds in extra symbols for debugging. -g alone can slow codes down by at least 2x. With the default flags, the Fortran code was the same speed as the Numpy version.

If we want efficient code, we need '-O3 -g0'.  If we test the code, and are confident in it, we can turn off debug symbols for the Fortran code only with '-g0' and get at least 2x speed up.  So, in this scenario and as is the case with this PR, the Fortran extension with -O3 and -g0 was 2x faster than the vectorized Numpy function, as well as having far less memory overhead since I used scalars in a double Fortran loop rather than array operations. Also, I initially wrote the Fortran code in vectorized notation, and the speed was the same as the version with loops so I did test some of these ideas.