@prisae 2018-09-01
Thanks for the further explanations @leonfoks, very interesting and insightful. As I said, `simpegEM1D` is the brainchild of @sgkang, I only contributed a negligible part to it, and never looked at the code in detail. So you'll have to wait for his response/opinion.