@sgkang 2018-09-26
Hi @leonfoks, sorry for being too late about this pull request!
I am just having some time to work on this, and excited with your improvements. I'll start to put some questions as I am reviewing this pull request!